refactor(admin): AdminMatieres sous 300 lignes — éclatement complet (#G1 ≤300)
AdminMatieres.vue : 776 -> 238. Composable useAdminMatieres (état + filtres +
regroupement par niveau/stats + cache + handlers modales) ; sous-composants
MatieresFilters (v-model recherche/filière/niveau) et MatieresTable avec leur CSS.
La vue orchestre header + états + filtres + table + MatiereModals (déjà extraite).
CSS distribué ; CSS mort (sélecteurs sans cible) retiré. Header bespoke conservé
verbatim (pas de @use admin-shared pour éviter un conflit de styles).

Tests : useAdminMatieres (8 cas, services mockés) + MatieresFilters + MatieresTable.
20/20 verts (avec MatiereModals non-régression). vite build vert. Fichiers ≤300.
